+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- OAI harvest cna now have wait times configured in the config files.
+- OAI harvest can now have wait times configured in the config files.
 
 ### Fixed
 - OAI jobs will also be marked as stopped if getting APIException  (dk.kb class)